requirements.txt
python-dotenv
openai==0.28.0
thefuzz
pandas